[PB-26] Move transaction generation out of send loop
Instead of placing pairs of (key, neighbours) into the queue, and then
constructing and sending transactions from that information in a loop,
we fill the queue with pairs of (tx, neighbours), and then only send
the transactions in the loop. If any part of transaction generation is
slow, it will no longer decrese the sending rate (only increase the
wait time at the start).